Create Codsoft_taskn03.py
This Python program is a menu-driven To-Do List application that runs in the command line. It allows the user to manage daily tasks efficiently by providing options to add, view, update, and delete tasks. The program stores all tasks in a list, where each task is saved as a text item. A function is used to display a menu that shows all available operations. The program runs inside a while loop, so the menu keeps appearing until the user chooses to exit